Selected shorter name for BTree.
