+++ src/Playground/InMemoryTreeBenchmark/RandomMidSizeInserts.cs
@@ -1,5 +1,5 @@
 ﻿using Tenray.ZoneTree.Collections;
-using Tenray.ZoneTree.Collections.BplusTree;
+using Tenray.ZoneTree.Collections.BTree;
 
 namespace Playground.InMemoryTreeBenchmark;
 
@@ -20,10 +20,10 @@ public static class RandomMidSizeInserts
         return arr;
     }
 
-    public static void InsertBplusTree(MidSizeKey[] arr)
+    public static void InsertBTree(MidSizeKey[] arr)
     {
         var count = arr.Length;
-        var tree = new BplusTreeUnsafe<MidSizeKey, MidSizeKey>(new MidSizeRefComparer());
+        var tree = new UnsafeBTree<MidSizeKey, MidSizeKey>(new MidSizeRefComparer());
         for(var i = 0; i < count; ++i)
         {
             var x = arr[i];
