+++ src/ZoneTree/Collections/BplusTree/BTree.cs
@@ -1,4 +1,4 @@
-﻿namespace Tenray.ZoneTree.Collections.BplusTree;
+﻿namespace Tenray.ZoneTree.Collections.BTree;
 
 /// <summary>
 /// In memory B+Tree.
@@ -6,7 +6,7 @@
 /// </summary>
 /// <typeparam name="TKey">Key Type</typeparam>
 /// <typeparam name="TValue">Value Type</typeparam>
-public partial class SafeBplusTree<TKey, TValue>
+public partial class BTree<TKey, TValue>
 {
     readonly NoLock TopLevelLocker = new();
 
@@ -26,7 +26,7 @@ public partial class SafeBplusTree<TKey, TValue>
 
     public int Length => _length;
 
-    public SafeBplusTree(
+    public BTree(
         IRefComparer<TKey> comparer,
         int nodeSize = 128,
         int leafSize = 128)
