+++ src/ZoneTree/Collections/BplusTree/UnsafeBTree.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 
-namespace Tenray.ZoneTree.Collections.BplusTree;
+namespace Tenray.ZoneTree.Collections.BTree;
 
 /// <summary>
 /// In memory B+Tree.
@@ -8,7 +8,7 @@ namespace Tenray.ZoneTree.Collections.BplusTree;
 /// </summary>
 /// <typeparam name="TKey">Key Type</typeparam>
 /// <typeparam name="TValue">Value Type</typeparam>
-public class BplusTreeUnsafe<TKey, TValue>
+public class UnsafeBTree<TKey, TValue>
 {
     readonly int NodeSize = 128;
 
@@ -28,7 +28,7 @@ public class BplusTreeUnsafe<TKey, TValue>
 
     public LeafNode Last => LastLeafNode;
 
-    public BplusTreeUnsafe(
+    public UnsafeBTree(
         IRefComparer<TKey> comparer,
         int nodeSize = 128,
         int leafSize = 128)
